+++ src/Bzn.Cloudios.Domain/Enums/ContainerStatus.cs
@@ -1,0 +1,9 @@
+namespace Bzn.Cloudios.Domain.Enums;
+
+public enum ContainerStatus
+{
+    Deploying,
+    Running,
+    Stopped,
+    Failed
+}